feat: Implement RepayForm component for loan repayment functionality
feat: Create SupplyCollateralModal for supplying collateral with health impact calculations

feat: Add WithdrawCollateralModal for withdrawing collateral with health impact assessment
